Move ContractValue to output/types
