+++ packages/discovery/src/blip/executeBlip.ts
@@ -1,5 +1,5 @@
 import { assert } from '@l2beat/shared-pure'
-import type { ContractValue } from '../discovery/utils/types'
+import type { ContractValue } from '../discovery/output/types'
 import type { BlipSexp } from './type'
 
 export function executeBlip(v: ContractValue, blip: BlipSexp): ContractValue {
