+++ packages/discovery/src/discovery/handlers/Handler.ts
@@ -1,6 +1,6 @@
 import type { EthereumAddress } from '@l2beat/shared-pure'
 import type { utils } from 'ethers'
-import type { ContractValue } from '../utils/types'
+import type { ContractValue } from '../output/types'
 
 import type { IProvider } from '../provider/IProvider'
