+++ packages/discovery/src/discovery/handlers/HandlerExecutor.ts
@@ -1,8 +1,8 @@
 import type { EthereumAddress } from '@l2beat/shared-pure'
 import type { ContractConfig } from '../config/ContractConfig'
 import type { DiscoveryCustomType } from '../config/RawDiscoveryConfig'
+import type { ContractValue } from '../output/types'
 import type { IProvider } from '../provider/IProvider'
-import type { ContractValue } from '../utils/types'
 import type { HandlerResult } from './Handler'
 import { decodeHandlerResults } from './decodeHandlerResults'
 import { executeHandlers } from './executeHandlers'
