+++ packages/discovery/src/discovery/handlers/reference.ts
@@ -1,5 +1,5 @@
 import * as z from 'zod'
-import type { ContractValue } from '../utils/types'
+import type { ContractValue } from '../output/types'
 
 import type { EthereumAddress } from '@l2beat/shared-pure'
 import type { IProvider } from '../provider/IProvider'
