+++ packages/discovery/src/discovery/handlers/user/AccessControlHandler.ts
@@ -2,8 +2,8 @@ import { assert, EthereumAddress } from '@l2beat/shared-pure'
 import { type providers, utils } from 'ethers'
 import * as z from 'zod'
 
+import type { ContractValue } from '../../output/types'
 import type { IProvider } from '../../provider/IProvider'
-import type { ContractValue } from '../../utils/types'
 import type { Handler, HandlerResult } from '../Handler'
 
 export type AccessControlHandlerDefinition = z.infer<
