+++ packages/discovery/src/discovery/handlers/user/ConstructorArgsHandler.ts
@@ -1,7 +1,7 @@
 import { assert, type EthereumAddress } from '@l2beat/shared-pure'
 import { ethers } from 'ethers'
 import { z } from 'zod'
-import type { ContractValue } from '../../utils/types'
+import type { ContractValue } from '../../output/types'
 
 import type { IProvider } from '../../provider/IProvider'
 import type { Handler, HandlerResult } from '../Handler'
