+++ packages/discovery/src/discovery/handlers/user/EventHandler.ts
@@ -7,9 +7,9 @@ import { groupBy } from 'lodash'
 import { executeBlip } from '../../../blip/executeBlip'
 import type { BlipSexp } from '../../../blip/type'
 import { validateBlip } from '../../../blip/validateBlip'
+import type { ContractValue } from '../../output/types'
 import { orderLogs } from '../../provider/BatchingAndCachingProvider'
 import type { IProvider } from '../../provider/IProvider'
-import type { ContractValue } from '../../utils/types'
 import type { Handler, HandlerResult } from '../Handler'
 import { getEventFragment } from '../utils/getEventFragment'
 import { toContractValue } from '../utils/toContractValue'
