+++ packages/discovery/src/discovery/handlers/user/StarkWareGovernanceHandler.ts
@@ -1,7 +1,7 @@
 import type { EthereumAddress } from '@l2beat/shared-pure'
 import { utils } from 'ethers'
 import * as z from 'zod'
-import type { ContractValue } from '../../utils/types'
+import type { ContractValue } from '../../output/types'
 
 import type { IProvider } from '../../provider/IProvider'
 import type { Handler, HandlerResult } from '../Handler'
