+++ packages/discovery/src/discovery/handlers/user/ZKsyncEraScheduledTransactionHandler.ts
@@ -3,10 +3,10 @@ import { utils } from 'ethers'
 import * as z from 'zod'
 
 import { uniqBy } from 'lodash'
+import type { ContractValue } from '../../output/types'
 import type { IProvider } from '../../provider/IProvider'
 import { ProxyDetector } from '../../proxies/ProxyDetector'
 import { get$Implementations } from '../../utils/extractors'
-import type { ContractValue } from '../../utils/types'
 import type { Handler, HandlerResult } from '../Handler'
 import { toContractValue } from '../utils/toContractValue'
