+++ packages/discovery/src/discovery/handlers/utils/bytes32ToContractValue.ts
@@ -1,5 +1,5 @@
 import { type Bytes, EthereumAddress } from '@l2beat/shared-pure'
-import type { ContractValue } from '../../utils/types'
+import type { ContractValue } from '../../output/types'
 
 export function bytes32ToContractValue(
   value: Bytes,
