+++ packages/discovery/src/discovery/handlers/utils/valueToBigInt.ts
@@ -1,5 +1,5 @@
 import { Bytes } from '@l2beat/shared-pure'
-import type { ContractValue } from '../../utils/types'
+import type { ContractValue } from '../../output/types'
 
 export function valueToBigInt(value: bigint | Bytes | ContractValue): bigint {
   if (value instanceof Bytes) {
