+++ packages/discovery/src/discovery/handlers/utils/valueToNumber.ts
@@ -1,5 +1,5 @@
 import type { Bytes } from '@l2beat/shared-pure'
-import type { ContractValue } from '../../utils/types'
+import type { ContractValue } from '../../output/types'
 
 import { valueToBigInt } from './valueToBigInt'
