+++ packages/discovery/src/discovery/proxies/auto/ArbitrumProxy.ts
@@ -5,7 +5,7 @@ import {
   Hash256,
   UnixTime,
 } from '@l2beat/shared-pure'
-import type { ContractValue } from '../../utils/types'
+import type { ContractValue } from '../../output/types'
 import type { ProxyDetails } from '../types'
 
 import { type providers, utils } from 'ethers'
