+++ packages/discovery/src/discovery/proxies/auto/StarkWareProxy.ts
@@ -6,7 +6,7 @@ import {
   UnixTime,
 } from '@l2beat/shared-pure'
 import { BigNumber, type providers, utils } from 'ethers'
-import type { ContractValue } from '../../utils/types'
+import type { ContractValue } from '../../output/types'
 import type { ProxyDetails } from '../types'
 
 import { type Semver, parseSemver } from '../../../utils/semver'
