+++ packages/discovery/src/discovery/type-casters/BaseTypeCaster.ts
@@ -1,4 +1,4 @@
-import type { ContractValue } from '../utils/types'
+import type { ContractValue } from '../output/types'
 
 export type ArgType = { [key: string]: string | number | boolean }
