+++ packages/discovery/src/discovery/type-casters/FormatSeconds.ts
@@ -1,5 +1,5 @@
 import { assert, formatSeconds } from '@l2beat/shared-pure'
-import type { ContractValue } from '../utils/types'
+import type { ContractValue } from '../output/types'
 import type { ArgType, BaseTypeCaster } from './BaseTypeCaster'
 
 export const FormatSeconds: BaseTypeCaster = {
