+++ packages/discovery/src/discovery/type-casters/GreaterThan.ts
@@ -2,7 +2,7 @@ import { assert } from '@l2beat/shared-pure'
 import { BigNumber } from 'bignumber.js'
 import { z } from 'zod'
 import { toContractValue } from '../handlers/utils/toContractValue'
-import type { ContractValue } from '../utils/types'
+import type { ContractValue } from '../output/types'
 import type { ArgType, BaseTypeCaster } from './BaseTypeCaster'
 
 const Validator = z.object({
