+++ packages/discovery/src/discovery/type-casters/Mapping.ts
@@ -1,6 +1,6 @@
 import { assert } from '@l2beat/shared-pure'
 import { z } from 'zod'
-import type { ContractValue } from '../utils/types'
+import type { ContractValue } from '../output/types'
 import type { ArgType, BaseTypeCaster } from './BaseTypeCaster'
 
 const Validator = z.record(
