+++ packages/discovery/src/discovery/type-casters/TypeApplier.ts
@@ -1,5 +1,5 @@
 import type { utils } from 'ethers'
-import type { ContractValue } from '../utils/types'
+import type { ContractValue } from '../output/types'
 
 import { assert } from '@l2beat/shared-pure'
 import type { DiscoveryCustomType } from '../config/RawDiscoveryConfig'
