--- packages/discovery/src/discovery/utils/types.ts
@@ -1,6 +1,0 @@
-export type ContractValue =
-  | string
-  | number
-  | boolean
-  | ContractValue[]
-  | { [key: string]: ContractValue | undefined }
