+++ packages/discovery/src/index.ts
@@ -77,6 +77,6 @@ export type {
   EoaParameters,
   ReceivedPermission,
   ResolvedPermissionPath,
+  ContractValue,
 } from './discovery/output/types'
-export type { ContractValue } from './discovery/utils/types'
 export { get$Implementations } from './discovery/utils/extractors'
